+++ packages/ricos-content-query/web/src/editor-query/editor-query-creator.ts
@@ -2,13 +2,20 @@ import type { Editor } from '@tiptap/core';
 import type { RicosStyles } from 'ricos-styles';
 import type { IEditorQuery } from 'ricos-types';
 import { EditorQuery } from './editor-query';
-import { activeDecorationsByTypeQuery, activeTextStylesQuery } from './queries';
+import {
+  activeDecorationsByTypeQuery,
+  activeTextStylesQuery,
+  isSingleRootTextBlock,
+  isMultipleSelection,
+} from './queries';
 
 const editorQueryCreator = (editor: Editor, ricosStyles: RicosStyles): IEditorQuery => {
   const editorQuery = new EditorQuery(editor, ricosStyles);
 
   editorQuery.addQuery(activeDecorationsByTypeQuery.NAME, activeDecorationsByTypeQuery.query);
   editorQuery.addQuery(activeTextStylesQuery.NAME, activeTextStylesQuery.query);
+  editorQuery.addQuery(isSingleRootTextBlock.NAME, isSingleRootTextBlock.query);
+  editorQuery.addQuery(isMultipleSelection.NAME, isMultipleSelection.query);
 
   editorQuery.query;
   return editorQuery;