+++ packages/ricos-content-query/web/src/editor-query/queries/index.ts
@@ -1,2 +1,4 @@
 export { default as activeDecorationsByTypeQuery } from './active-decoration-by-type';
 export { default as activeTextStylesQuery } from './active-text-styles';
+export { default as isSingleRootTextBlock } from './is-single-root-textblock';
+export { default as isMultipleSelection } from './is-multiple-selection';