+++ packages/ricos-context/web/src/ToolbarContext.tsx
@@ -10,6 +10,7 @@ import type {
   RicosPortal,
   AddLinkData,
   IRicosToolbars,
+  IEditorQuery,
 } from 'ricos-types';
 import type { LinkSettings, ToolbarSettings, RicosCssOverride } from 'ricos-common';
 import { RicosContext } from './RicosContext';
@@ -52,6 +53,7 @@ export type ToolbarContextType = {
     TEXT_HIGHLIGHT: Record<string, any>;
   };
   defaultLineSpacing: { 'line-height'?: string; 'padding-top'?: string; 'padding-bottom'?: string };
+  editorQuery: IEditorQuery;
 };
 
 export const ToolbarContext = React.createContext<ToolbarContextType>({
@@ -72,6 +74,7 @@ export const ToolbarContext = React.createContext<ToolbarContextType>({
   toolbars: null as unknown as IRicosToolbars,
   colorPickerData: {} as unknown as ToolbarContextType['colorPickerData'],
   defaultLineSpacing: {},
+  editorQuery: null as unknown as IEditorQuery,
 });
 
 export const withToolbarContext = WrappedComponent => {