+++ packages/ricos-plugins/web/src/editorPlugins.ts
@@ -13,7 +13,10 @@ import { RicosPluginAddButtons } from './plugin-add-buttons';
 import { mergeConfig } from './mergeConfig';
 import { EditorPlugin } from './editorPlugin';
 
-export type PluginServices = Omit<RicosServices, 'plugins' | 'editor' | 'zIndexService'>;
+export type PluginServices = Omit<
+  RicosServices,
+  'plugins' | 'editor' | 'zIndexService' | 'editorQuery'
+>;
 
 export class PluginCollisionError extends Error {}